backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic[email]==2.9.2
numpy==2.1.1
scikit-learn==1.5.2
celery==5.4.0
redis==5.0.8
httpx==0.27.2
pytest==8.3.3
sqlalchemy==2.0.25
sqlalchemy-utils==0.41.1
requests==2.32.3
psycopg2-binary==2.9.9
